requirements.txt
django
psycopg2
djangorestframework
 django-cors-headers==3.5.0
